Refactor: split general digits calculation from single digit function
