+++ calc.go
@@ -5,29 +5,11 @@ package pi
 // Based on source provided by Fabrice Bellard, taken from https://bellard.org/pi/pi.c
 
 import (
-	"context"
 	"fmt"
 	"math"
 
 	"go.uber.org/zap"
 )
-
-var (
-	logger       = zap.NewNop()
-	cache  Cache = NewNoopCache()
-)
-
-func SetLogger(l *zap.Logger) {
-	if l != nil {
-		logger = l
-	}
-}
-
-func SetCache(c Cache) {
-	if c != nil {
-		cache = c
-	}
-}
 
 // Returns the inverse of x mod y
 func invMod(x int64, y int64) int64 {
@@ -203,35 +185,4 @@ func CalcDigits(n uint64) string {
 		zap.String("result", result),
 	)
 	return result
-}
-
-func PiDigit(ctx context.Context, n uint64) (string, error) {
-	l := logger.With(
-		zap.Uint64("n", n),
-	)
-	l.Debug("PiDigits: enter")
-	index := uint64(n/9) * 9
-	key := fmt.Sprintf("%d", index)
-	digits, err := cache.GetValue(ctx, key)
-	if err != nil {
-		logger.Error("Error retrieving digits from cache",
-			zap.Error(err),
-		)
-		return "", err
-	}
-	if digits == "" {
-		digits = CalcDigits(index)
-		err = cache.SetValue(ctx, key, digits)
-		if err != nil {
-			logger.Error("Error writing digits to cache",
-				zap.Error(err),
-			)
-			return "", err
-		}
-	}
-	digit := string(digits[n%9])
-	logger.Debug("PiDigit: exit",
-		zap.String("digit", digit),
-	)
-	return digit, nil
 }
